solving comments of PR
