+++ src/components/PlayButton/PlayButton.module.scss
@@ -61,9 +61,6 @@ $icon-size: 0.7rem;
     color: colors.getColor("neutral", 60);
     background-color: colors.getColor("neutral", 90);
 
-    //box-shadow: none;
-    //transform: translateY(0);
-
     &:hover {
         color: colors.getColor("fault", 70);
         background-color: colors.getColor("fault", 90);
