+++ src/pages/TestingPage/TestControls/ControlButtons/ControlButtons.module.scss
@@ -1,0 +1,5 @@
+.controlsWrapper {
+    display: flex;
+    flex-flow: row;
+    justify-content: space-evenly;
+}
