+++ src/pages/TestingPage/TestControls/TestAttributes.ts
@@ -10,3 +10,34 @@ export type ToggleInputAttributes = {
 };
 
 export type TestAttributes = Record<string, ToggleInputAttributes>;
+
+export type InputData = {
+    id: string;
+    type: string;
+    value: number | null;
+    enabled: boolean;
+    validity: { isValid: boolean; msg: string };
+};
+
+export type ChangingValue = {
+    id: string;
+    value: number | null;
+};
+
+export type EnablingValue = {
+    id: string;
+    enabled: boolean;
+};
+
+export type FormData = Array<InputData>;
+
+export type Form = { formData: FormData; isValid: boolean };
+
+export type SubmitHandler = () => void;
+export type ChangeValue = (id: string, value: number) => void;
+export type ChangeEnable = (id: string, enable: boolean) => void;
+
+export type Action =
+    | { type: "CHANGE VALUE"; payload: ChangingValue }
+    | { type: "CHANGE ENABLE"; payload: EnablingValue }
+    | { type: "RESET INITIAL STATE"; payload: Form };
