+++ src/pages/TestingPage/TestControls/TestControls.module.scss
@@ -32,21 +32,6 @@
     font-size: large;
 }
 
-.controlsWrapper {
-    display: flex;
-    flex-flow: row;
-    justify-content: space-evenly;
-}
-
-.inputWrapper {
-    display: flex;
-    flex-flow: column;
-    overflow-y: scroll;
-    flex: 1 1 0;
-    min-height: 0;
-    gap: 0.2rem;
-}
-
 .body {
     display: flex;
     flex-direction: column;
